Dockerfiles: switch to alpine to add libcap
When using index-headers in thanos store, store may panic due to
reaching the "ulimit -l". The error and documentation says tuning the
ulimits should or could be done.

As long as busybox is used as base image we are (afaik) not able to set
the capabilities to set other values for "ulimit -l" when using the
upstream images.